Convert an object to a simple feature or bounding box object

Both functions will pass a NULL value without returning an error. If a POINT or MULTIPOINT object is passed to as_bbox() a 0.00000001 meter buffer is applied. If a character object is passed to as_bbox() it is passed to osmdata::getbb() using format_out = "matrix" which is converted into a bounding box.

Usage

as_sf(x, crs = NULL, sf_col = "geometry", ext = TRUE, ...)

as_bbox(x, crs = NULL, ext = TRUE, ...)

as_sfc(x, crs = NULL, ext = TRUE, ...)

as_sf_list(x, nm = "data", col = NULL, crs = NULL, clean_names = TRUE)

as_sf_class(x, class = NULL, allow_null = TRUE, call = caller_env(), ...)

as_crs(crs = NULL, check = FALSE, call = parent.frame())

as_wgs84(x)

Arguments

x

A sf, bbox, sfc, raster, sp, or data frame object that can be converted into a simple feature or bounding box object. as_bbox() can also convert a vector with xmin, ymin, xmax, and ymax values. as_sf_list() only supports sf objects or a data frames with a sf list column named "data" (typically created by using dplyr::group_nest() on an sf object.

crs

Coordinate reference system for sf, bbox, sfc or sf list object to return.

sf_col

A column name to use for the geometry column created by as_sf; defaults to "geometry".

ext

If TRUE, as_sf will convert a data frame or character vector of addresses to an sf object using df_to_sf or address_to_sf. If FALSE, only spatial objects (bbox, sfg, sfc, sf list, raster, or sp objects) can be converted. Defaults to TRUE.

...

Additional parameters passed to as_sf, as_sfc, as_bbox, or as_sf_list

nm

For as_sf_list, name(s) for sf list; defaults to "data". If col is provided, the values of the grouping column are used as names.

col

For as_sf_list, the name of the column used to group data if x is a sf object or used to group and nest data before passing to x.

clean_names

If TRUE, clean names provided to nm or created based on value of col using janitor::clean_names. If FALSE, use names as provided.

class

A class to convert data to; defaults to NULL (which returns "sf")

allow_null

For as_sf_class, if class is NULL and allow_null is TRUE, return x without any class conversion or checks. Defaults to TRUE.

call

Passed as the call parameter for cli::cli_abort or rlang::arg_match to improve error messages when function is used internally.

check

For as_crs(), if TRUE, error if crs cannot be converted to a valid coordinate reference system. Defaults to FALSE.

data

Data that can be converted to sf, sfc, bbox or a sf list object.

Examples

nc <- read_sf_ext(system.file("shape/nc.shp", package = "sf"))
nc[["category"]] <- sample(c("A", "B", "C", "D", "E"), nrow(nc), replace = TRUE)

as_sf(nc$geometry)
#> Simple feature collection with 100 features and 0 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
#> First 10 features:
#>                          geometry
#> 1  MULTIPOLYGON (((-81.47276 3...
#> 2  MULTIPOLYGON (((-81.23989 3...
#> 3  MULTIPOLYGON (((-80.45634 3...
#> 4  MULTIPOLYGON (((-76.00897 3...
#> 5  MULTIPOLYGON (((-77.21767 3...
#> 6  MULTIPOLYGON (((-76.74506 3...
#> 7  MULTIPOLYGON (((-76.00897 3...
#> 8  MULTIPOLYGON (((-76.56251 3...
#> 9  MULTIPOLYGON (((-78.30876 3...
#> 10 MULTIPOLYGON (((-80.02567 3...

as_bbox(nc)
#>      xmin      ymin      xmax      ymax 
#> -84.32385  33.88199 -75.45698  36.58965 

as_sf_list(nc, col = "category")
#> $a
#> Simple feature collection with 25 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -83.25912 ymin: 34.36442 xmax: -76.35819 ymax: 36.55629
#> Geodetic CRS:  NAD27
#> # A tibble: 25 × 16
#>     AREA PERIMETER CNTY_ CNTY_ID NAME  FIPS  FIPSNO CRESS_ID BIR74 SID74 NWBIR74
#>    <dbl>     <dbl> <dbl>   <dbl> <chr> <chr>  <dbl>    <int> <dbl> <dbl>   <dbl>
#>  1 0.097      1.67  1833    1833 Hert… 37091  37091       46  1452     7     954
#>  2 0.114      1.35  1838    1838 Casw… 37033  37033       17  1035     2     550
#>  3 0.109      1.32  1841    1841 Pers… 37145  37145       73  1556     4     613
#>  4 0.128      1.55  1897    1897 Fran… 37069  37069       35  1399     2     736
#>  5 0.108      1.48  1900    1900 Fors… 37067  37067       34 11858    10    3919
#>  6 0.131      1.52  1928    1928 Edge… 37065  37065       33  3657    10    2186
#>  7 0.08       1.31  1936    1936 Yanc… 37199  37199      100   770     0      12
#>  8 0.066      1.07  1950    1950 Alex… 37003  37003        2  1333     0     128
#>  9 0.1        1.33  1962    1962 Wash… 37187  37187       94   990     5     521
#> 10 0.116      1.66  1964    1964 McDo… 37111  37111       56  1946     5     134
#> # ℹ 15 more rows
#> # ℹ 5 more variables: BIR79 <dbl>, SID79 <dbl>, NWBIR79 <dbl>,
#> #   geometry <MULTIPOLYGON [°]>, category <chr>
#> 
#> $b
#> Simple feature collection with 20 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.03086 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
#> # A tibble: 20 × 16
#>     AREA PERIMETER CNTY_ CNTY_ID NAME  FIPS  FIPSNO CRESS_ID BIR74 SID74 NWBIR74
#>    <dbl>     <dbl> <dbl>   <dbl> <chr> <chr>  <dbl>    <int> <dbl> <dbl>   <dbl>
#>  1 0.114      1.44  1825    1825 Ashe  37009  37009        5  1091     1      10
#>  2 0.07       2.97  1831    1831 Curr… 37053  37053       27   508     1     123
#>  3 0.124      1.43  1837    1837 Stok… 37169  37169       85  1612     1     160
#>  4 0.143      1.66  1840    1840 Gran… 37077  37077       39  1671     4     930
#>  5 0.072      1.08  1842    1842 Vance 37181  37181       91  2180     4    1179
#>  6 0.199      1.98  1874    1874 Wilk… 37193  37193       97  3146     4     200
#>  7 0.063      1     1881    1881 Perq… 37143  37143       72   484     1     230
#>  8 0.064      1.21  1892    1892 Avery 37011  37011        6   781     0       4
#>  9 0.086      1.27  1893    1893 Yadk… 37197  37197       99  1269     1      65
#> 10 0.077      1.27  1908    1908 Durh… 37063  37063       32  7970    16    3732
#> 11 0.094      3.64  2000    2000 Dare  37055  37055       28   521     0      43
#> 12 0.181      1.98  2040    2040 Moore 37125  37125       63  2648     5     844
#> 13 0.078      1.20  2056    2056 Grah… 37075  37075       38   415     0      40
#> 14 0.131      1.68  2082    2082 Macon 37113  37113       57   797     0       9
#> 15 0.241      2.21  2083    2083 Samp… 37163  37163       82  3025     4    1396
#> 16 0.121      1.98  2091    2091 Jones 37103  37103       52   578     1     297
#> 17 0.167      2.71  2099    2099 Hyde  37095  37095       48   338     0     134
#> 18 0.177      2.92  2119    2119 Crav… 37049  37049       25  5868    13    1744
#> 19 0.125      2.87  2156    2156 Cart… 37031  37031       16  2414     5     341
#> 20 0.212      2.02  2241    2241 Brun… 37019  37019       10  2181     5     659
#> # ℹ 5 more variables: BIR79 <dbl>, SID79 <dbl>, NWBIR79 <dbl>,
#> #   geometry <MULTIPOLYGON [°]>, category <chr>
#> 
#> $c
#> Simple feature collection with 17 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.94867 xmax: -77.07531 ymax: 36.56521
#> Geodetic CRS:  NAD27
#> # A tibble: 17 × 16
#>     AREA PERIMETER CNTY_ CNTY_ID NAME  FIPS  FIPSNO CRESS_ID BIR74 SID74 NWBIR74
#>    <dbl>     <dbl> <dbl>   <dbl> <chr> <chr>  <dbl>    <int> <dbl> <dbl>   <dbl>
#>  1 0.143      1.63  1828    1828 Surry 37171  37171       86  3188     5     208
#>  2 0.153      2.21  1832    1832 Nort… 37131  37131       66  1421     9    1066
#>  3 0.17       1.68  1903    1903 Guil… 37081  37081       41 16184    23    5483
#>  4 0.111      1.39  1904    1904 Alam… 37001  37001        1  4672    13    1243
#>  5 0.104      1.29  1907    1907 Oran… 37135  37135       68  3164     4     776
#>  6 0.118      1.60  1946    1946 Madi… 37115  37115       58   765     2       5
#>  7 0.134      1.59  1980    1980 Rowan 37159  37159       80  4606     3    1057
#>  8 0.141      2.32  2013    2013 Swain 37173  37173       87   675     3     281
#>  9 0.065      1.09  2026    2026 Lee   37105  37105       53  2252     5     736
#> 10 0.142      1.66  2029    2029 Wayne 37191  37191       96  6638    18    2593
#> 11 0.154      1.68  2030    2030 Harn… 37085  37085       43  3776     6    1051
#> 12 0.118      1.51  2032    2032 Clev… 37045  37045       23  4866    10    1491
#> 13 0.13       1.73  2044    2044 Mont… 37123  37123       62  1258     3     472
#> 14 0.12       1.69  2088    2088 Cher… 37039  37039       20  1027     2      32
#> 15 0.172      1.84  2090    2090 Cumb… 37051  37051       26 20366    38    7043
#> 16 0.24       2.00  2150    2150 Robe… 37155  37155       78  7889    31    5904
#> 17 0.24       2.37  2232    2232 Colu… 37047  37047       24  3350    15    1431
#> # ℹ 5 more variables: BIR79 <dbl>, SID79 <dbl>, NWBIR79 <dbl>,
#> #   geometry <MULTIPOLYGON [°]>, category <chr>
#> 
#> $d
#> Simple feature collection with 18 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -83.36472 ymin: 34.06203 xmax: -76.11271 ymax: 36.54789
#> Geodetic CRS:  NAD27
#> # A tibble: 18 × 16
#>     AREA PERIMETER CNTY_ CNTY_ID NAME  FIPS  FIPSNO CRESS_ID BIR74 SID74 NWBIR74
#>    <dbl>     <dbl> <dbl>   <dbl> <chr> <chr>  <dbl>    <int> <dbl> <dbl>   <dbl>
#>  1 0.153     1.62   1839    1839 Rock… 37157  37157       79  4449    16    1243
#>  2 0.053     1.17   1848    1848 Pasq… 37139  37139       70  1638     3     622
#>  3 0.081     1.29   1880    1880 Wata… 37189  37189       95  1323     1      17
#>  4 0.044     1.16   1887    1887 Chow… 37041  37041       21   751     1     368
#>  5 0.18      2.15   1905    1905 Bert… 37015  37015        8  1324     6     921
#>  6 0.118     1.90   1937    1937 Mart… 37117  37117       59  1549     2     883
#>  7 0.145     1.79   1951    1951 Davi… 37057  37057       29  5509     8     736
#>  8 0.201     1.80   1968    1968 Rand… 37151  37151       76  4456     7     384
#>  9 0.168     1.79   1984    1984 Pitt  37147  37147       74  5094    14    2620
#> 10 0.203     3.20   2004    2004 Beau… 37013  37013        7  2692     7    1131
#> 11 0.125     1.60   2039    2039 Jack… 37099  37099       50  1143     2     215
#> 12 0.143     1.89   2041    2041 Meck… 37119  37119       60 21588    44    8027
#> 13 0.104     1.55   2065    2065 Leno… 37107  37107       54  3589    10    1826
#> 14 0.06      1.04   2071    2071 Polk  37149  37149       75   533     1      95
#> 15 0.082     1.39   2085    2085 Paml… 37137  37137       69   542     1     222
#> 16 0.098     1.26   2097    2097 Hoke  37093  37093       47  1494     7     987
#> 17 0.121     1.86   2107    2107 Rich… 37153  37153       77  2756     4    1043
#> 18 0.042     0.999  2238    2238 New … 37129  37129       65  5526    12    1633
#> # ℹ 5 more variables: BIR79 <dbl>, SID79 <dbl>, NWBIR79 <dbl>,
#> #   geometry <MULTIPOLYGON [°]>, category <chr>
#> 
#> $e
#> Simple feature collection with 20 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -83.98855 ymin: 34.30505 xmax: -75.95718 ymax: 36.57286
#> Geodetic CRS:  NAD27
#> # A tibble: 20 × 16
#>     AREA PERIMETER CNTY_ CNTY_ID NAME  FIPS  FIPSNO CRESS_ID BIR74 SID74 NWBIR74
#>    <dbl>     <dbl> <dbl>   <dbl> <chr> <chr>  <dbl>    <int> <dbl> <dbl>   <dbl>
#>  1 0.061      1.23  1827    1827 Alle… 37005  37005        3   487     0      10
#>  2 0.062      1.55  1834    1834 Camd… 37029  37029       15   286     0     115
#>  3 0.091      1.28  1835    1835 Gates 37073  37073       37   420     0     254
#>  4 0.118      1.42  1836    1836 Warr… 37185  37185       93   968     4     748
#>  5 0.19       2.20  1846    1846 Hali… 37083  37083       42  3608    18    2365
#>  6 0.142      1.64  1913    1913 Nash  37127  37127       64  4021     8    1851
#>  7 0.059      1.32  1927    1927 Mitc… 37121  37121       61   671     0       1
#>  8 0.122      1.52  1932    1932 Cald… 37027  37027       14  3609     6     309
#>  9 0.219      2.13  1938    1938 Wake  37183  37183       92 14484    16    4397
#> 10 0.155      1.78  1947    1947 Ired… 37097  37097       49  4139     4    1144
#> 11 0.069      1.20  1948    1948 Davie 37059  37059       30  1207     1     148
#> 12 0.134      1.76  1958    1958 Burke 37023  37023       12  3573     5     326
#> 13 0.099      1.41  1963    1963 Tyrr… 37177  37177       89   248     0     116
#> 14 0.168      2.00  1988    1988 Bunc… 37021  37021       11  7515     9     930
#> 15 0.207      1.85  1989    1989 John… 37101  37101       51  3999     6    1165
#> 16 0.095      1.47  2047    2047 Hend… 37089  37089       45  2574     5     158
#> 17 0.098      1.39  2067    2067 Tran… 37175  37175       88  1173     3      92
#> 18 0.163      1.72  2095    2095 Union 37179  37179       90  3915     4    1034
#> 19 0.051      1.10  2109    2109 Clay  37043  37043       22   284     0       1
#> 20 0.214      2.15  2185    2185 Pend… 37141  37141       71  1228     4     580
#> # ℹ 5 more variables: BIR79 <dbl>, SID79 <dbl>, NWBIR79 <dbl>,
#> #   geometry <MULTIPOLYGON [°]>, category <chr>
#> 

as_crs(nc)
#> Coordinate Reference System:
#>   User input: NAD27 
#>   wkt:
#> GEOGCRS["NAD27",
#>     DATUM["North American Datum 1927",
#>         ELLIPSOID["Clarke 1866",6378206.4,294.978698213898,
#>             LENGTHUNIT["metre",1]]],
#>     PRIMEM["Greenwich",0,
#>         ANGLEUNIT["degree",0.0174532925199433]],
#>     CS[ellipsoidal,2],
#>         AXIS["latitude",north,
#>             ORDER[1],
#>             ANGLEUNIT["degree",0.0174532925199433]],
#>         AXIS["longitude",east,
#>             ORDER[2],
#>             ANGLEUNIT["degree",0.0174532925199433]],
#>     ID["EPSG",4267]]

as_xy(nc[1,])
#>           x       y
#> 1 -81.49823 36.4314