Bind columns based on sf objects:

bind_boundary_col uses sf::st_join to assign simple feature data to an enclosing polygon. This function with other join functions to add columns based on other spatial relationships, e.g. bind a column for the nearest point feature.

Usage

bind_boundary_col(x, boundary = NULL, join = NULL, col = "name", ...)

bind_units_col(
  x,
  y,
  units = NULL,
  drop = FALSE,
  keep_all = TRUE,
  .id = NULL,
  call = caller_env()
)

Arguments

boundary

An sf object with a column named "name" or a list of sf objects where all items in the list have a "name" column.

join

geometry predicate function; defaults to NULL, set to sf::st_intersects if key_list contains only POLYGON or MULTIPOLYGON objects or sf::st_nearest_feature if key_list contains other types.

col

Column name used to convert boundary into a sf list and contains the values that should be added as a column to the input sf object.

y

Vector of numeric or units values to bind to x.

units

Units to use for y (if numeric) or convert to (if y is units class); defaults to NULL.

drop

If TRUE, apply the units::drop_units function to the column with units class values and return numeric values instead; defaults to FALSE.

keep_all

If FALSE, keep all columns. If FALSE, return only the named .id column.

.id

Name to use for vector of units provided to "y" parameter, when "y" is bound to the "x" data frame or tibble as a new column.