Erase an area of a simple feature object

This function extends sf::st_difference() by unioning the second parameter by default, checking validity of inputs, and optionally (when flip = TRUE) using sf::st_intersection() instead of sf::st_difference. st_trim() is equivalent to st_erase() with flip set to TRUE.

Usage

st_erase(x, y, flip = FALSE, union = TRUE, ...)

st_trim(x, y, union = TRUE, ...)

Arguments

x

A sf, sfc, or bbox object to erase or trim.

y

A sf, sfc, or bbox object to use to erase or trim.

flip

If TRUE, use st_intersection; if FALSE use st_difference, Default: FALSE

union

If TRUE, use sf::st_combine and sf::st_union on y before applying difference/intersection; defaults to FALSE.

...

arguments passed on to s2_options