Make a grid over a location bounding box

Create a grid with an id column and optionally a set number of columns and rows. This documentation is incomplete the function may change.

Usage

st_make_grid_ext(
  x,
  crs = NULL,
  cols = NULL,
  rows = NULL,
  gutter = 0,
  desc = FALSE,
  n = NULL,
  cellsize = NULL,
  what = NULL,
  style = "rect",
  .id = "id"
)

Arguments

x

An object sf, bbox, sfc, raster, or sp object or a data frame that can be converted to an sf object or a list of sf, bbox, or sfc objects. st_bbox_asp also supports vectors in the same format as a bbox object.

crs

Coordinate reference system of bounding box to return

cols, rows

Used to set n if either are not NULL; defaults to NULL. row and id are added as columns to the grid if they are provided.

gutter

Distance in units between each column cell; gutter effectively serves as a margin as the negative buffer is applied to all cells (including those at the edges of the grid).

desc

If TRUE, reverse standard order of cell id numbering; defaults FALSE

n

If n is NULL and square is TRUE, the grid is set automatically to be 10 cells wide, Default: NULL

cellsize

target cellsize

what

"polygons", "corners", "centers"; set to centers automatically if style is "circle", "circle_offset" but a buffer is applied to return circular polygons.

style

Style of cell to return with options including "rect", "square", "hex", "flat_top_hex", "circle", "circle_offset"

.id

A name to use for the cell id column.

location

A sf, sfc, or bbox object, Default: NULL. Required.

See also

Examples