Nudge a simple feature to the center of another feature and/or a set distance

Nudge, move, or shift a sf, sfc, or bbox object to the center of another feature and/or by a set distance.

Usage

st_nudge(
  x,
  to = NULL,
  nudge_y = 0,
  nudge_x = 0,
  unit = NULL,
  scale = 1,
  rotate = 0,
  crs = NULL
)

Arguments

x

A sf, bbox, sfc, raster, sp, or data frame object that can be converted into a simple feature or bounding box object. as_bbox() can also convert a vector with xmin, ymin, xmax, and ymax values. as_sf_list() only supports sf objects or a data frames with a sf list column named "data" (typically created by using dplyr::group_nest() on an sf object.

to

sf object to use as new center for x or length 2 numeric vector with the nudge_y and nudge_x distance (in that order).

nudge_y, nudge_x

Distance to nudge geometry in unit. If unit is NULL, distance is assumed to be in the same units as the coordinate reference system of the input object.

unit

Units for nudge_y and nudge_x distance (also used if to is numeric).

scale

numeric; scale factor, Default: 1

rotate

numeric; degrees to rotate (-360 to 360), Default: 0

crs

Coordinate reference system for sf, bbox, sfc or sf list object to return.