Write an sf object to an svg file

write_sf_svg() uses plot() and svg() to create a simple plot of an sf object geometry. This function is convenient for working with designers or other collaborators interested in using spatial data outside of R or a desktop GIS application.

Usage

write_sf_svg(data, filename = NULL, path = NULL, ..., width = 10, height = 10)

Arguments

data

A sf object to save as a svg file.

filename

the file path of the output file(s). The page number is substituted if a C integer format is included in the character string, as in the default. (Depending on the platform, the result must be less than PATH_MAX characters long, and may be truncated if not. See pdf for further details.) Tilde expansion is performed where supported by the platform.

path

File path to combine with filename. Optional if filename is provided. filename is optional if path includes a svg file extension.

...

Arguments passed on to grDevices::svg

pointsize

the default pointsize of plotted text (in big points).

onefile

should all plots appear in one file or in separate files?

family

one of the device-independent font families, "sans", "serif" and "mono", or a character string specify a font family to be searched for in a system-dependent way.

On unix-alikes (incl.\ macOS), see the ‘Cairo fonts’ section in the help for X11.

bg

the initial background colour: can be overridden by setting par("bg").

antialias

string, the type of anti-aliasing (if any) to be used; defaults to "default".

symbolfamily

a length-one character string that specifies the font family to be used as the "symbol" font (e.g., for plotmath output).

width

the width of the device in inches.

height

the height of the device in inches.