Check if x is an sf object

If x is an sf object invisibly return TRUE. If not, return an error with cli::cli_abort

Usage

check_sf(
  x,
  arg = caller_arg(x),
  null.ok = FALSE,
  list.ok = FALSE,
  ext = FALSE,
  call = caller_env(),
  ...
)

Arguments

x

An sf, sfc, or bbox object.

null.ok

If TRUE and x is NULL, return TRUE; defaults to FALSE.

list.ok

If TRUE, return TRUE if x is an sf list or, if ext is TRUE, a list of sf, sfc, or bbox objects. Defaults to FALSE.

ext

If TRUE, check if x is a sf, sfc, or bbox class object or not; defaults to FALSE. (used by is_sf)

call

The execution environment of a currently running function, e.g. call = caller_env(). The corresponding function call is retrieved and mentioned in error messages as the source of the error.

You only need to supply call when throwing a condition from a helper function which wouldn't be relevant to mention in the message.

Can also be NULL or a defused function call to respectively not display any call or hard-code a code to display.

For more information about error calls, see Including function calls in error messages.

...

Arguments passed on to cli::cli_abort

message

It is formatted via a call to cli_bullets().

.envir

Environment to evaluate the glue expressions in.