Use mapview to interactively explore spatial data

A wrapper for mapview::mapview() that drops list columns and makes it easier to quickly specify a zcol value.

Usage

mapview_ext(x, zcol = NULL, remove_na = FALSE, ...)

mapview_exif(
  path = NULL,
  fileext = "jpeg",
  popup = TRUE,
  tooltip = FALSE,
  images = NULL,
  width = 320,
  ...
)

mapview_popup_img(
  images,
  popup = TRUE,
  tooltip = FALSE,
  map = NULL,
  width = 320,
  ...,
  call = caller_env()
)

Arguments

x

a Raster* or Spatial* or Satellite or sf or stars object or a list of any combination of those. Furthermore, this can also be a data.frame, a numeric vector or a character string pointing to a tile image folder or file on disk. If missing, a blank map will be drawn. A value of NULL will return NULL.

zcol

attribute name(s) or column number(s) in attribute table of the column(s) to be rendered. See also Details.

remove_na

If TRUE and zcol is not NULL, filter NA values from the data frame column defined by zcol before passing to mapview::mapview()

...

Arguments passed on to mapview::mapview

path

A path to folder or file.

fileext

File extension. Defaults to "jpeg".

popup

If TRUE, add a popup image to a leaflet map; defaults TRUE.

tooltip

logical, whether to show image(s) as popup(s) (on click) or tooltip(s) (on hover).

images

A simple feature object with columns for the image path/url, image width, and image height.

width

the width of the image(s) in pixels.

map

an optional existing map to be updated/added to.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

See also