Read EXIF metadata to create a simple feature object or write EXIF metadata to image files

read_sf_exif() read EXIF data from folder of files and, geometry is TRUE and coordinate metadata is available, convert the data to a sf object. This function also assigns a cardinal direction based on the direction metadata and recodes the orientation metadata.

For write_exif() the parameters are used to multiple tags with the same values:

title: Title, IPTC:Headline, IPTC:ObjectName, XMP-dc:Title

description: ImageDescription, XMP-dc:Description, and IPTC:Caption-Abstract

keywords: Keywords, IPTC:Keywords, XMP-dc:Subject

Usage

read_sf_exif(
  path = NULL,
  filetype = NULL,
  bbox = NULL,
  sort = NULL,
  tags = NULL,
  geometry = TRUE,
  ...
)

write_exif(
  path,
  filetype = NULL,
  title = NULL,
  author = NULL,
  date = NULL,
  keywords = NULL,
  description = NULL,
  args = NULL,
  overwrite = TRUE,
  append_keywords = FALSE
)

write_exif_from(
  path,
  filetype = NULL,
  from,
  .id = "name",
  tag = "keywords",
  join = NULL,
  overwrite = TRUE
)

Arguments

path

A path to folder or file.

filetype

The file extension or file type; defaults to NULL.

bbox

Optional bounding box to crop returned file (excluding images with location data outside the bounding box). If bbox is provided the returned data will match the crs of the bbox.

sort

Column name for variable to sort by passed to sort_features(). Currently supports "lon", "lat", or "filename". Defaults to NULL.

tags

Optional list of EXIF tags to read from files. Must include GPS tags to create an sf object.

geometry

If TRUE (defualt), return a simple feature object. If FALSE, return a data.frame.

...

Additional EXIF tags to pass to exiftoolr::exif_read

title

Title to add to file metadata with exiftoolr, Default: NULL.

author

Author to add to file metadata with exiftoolr, Default: NULL.

date

Date to add to file metadata with exiftoolr (not currently working), Default: NULL.

keywords

Keyword(s) added to file metadata with with exiftoolr, Default: NULL.

args

Alternate arguments passed to exiftoolr::exif_call(). If args is not NULL, title, author, date, and keywords are ignored; defaults to NULL.

overwrite

If TRUE, overwrite any existing EXIF metadata present in the provided fields; defaults to TRUE

from

A sf object or list of sf objects where each object has a column with a name matching the .id parameter. The attribute value in this column are used to assign the tag parameter to the file at the provided path based on the spatial relationship set by join. For example, from may be boundary data used to assign keywords based on photo locations.

.id

Column name in from with the values to use for tag values.

tag

EXIF tag to update, supported options include "keywords", "title", or "description".

join

geometry predicate function; defaults to NULL, set to sf::st_intersects if from contains only POLYGON or MULTIPOLYGON objects or sf::st_nearest_feature if from contains other types.

See also

Examples

if (FALSE) {
  read_sf_exif(
    path = system.file("extdata/photos", package = "overedge"),
    filetype = "jpeg"
  )
}