Read spatial data in a bounding box to a simple feature object from multiple sources

An extended version of sf::read_sf() that support reading spatial data based on a file path, URL, or the data name and associated package. Optionally provide a bounding box to filter data (not supported for all data types).

Usage

read_sf_ext(...)

read_sf_pkg(
  data,
  path = NULL,
  bbox = NULL,
  package = NULL,
  filetype = "gpkg",
  ...
)

read_sf_path(path, bbox = NULL, ...)

read_sf_query(
  path,
  dsn = NULL,
  bbox = NULL,
  query = NULL,
  table = NULL,
  name = NULL,
  name_col = NULL,
  wkt_filter = NULL,
  zm_drop = FALSE,
  ...
)

read_sf_excel(
  path,
  sheet = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  ...
)

read_sf_csv(
  path,
  url = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  show_col_types = FALSE,
  ...
)

read_sf_url(url, bbox = NULL, coords = NULL, zm_drop = TRUE, ...)

read_sf_esri(
  url,
  bbox = NULL,
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  ...
)

read_sf_gist(url, id = NULL, bbox = NULL, ...)

read_sf_gmap(url, bbox = NULL, zm_drop = TRUE)

read_sf_download(
  url,
  filename,
  bbox = NULL,
  path = NULL,
  filetype = "geojson",
  prefix = "date",
  method = "auto",
  unzip = FALSE,
  ...
)

read_sf_gsheet(
  url,
  sheet = NULL,
  ss = NULL,
  bbox = NULL,
  ask = FALSE,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  ...
)

Arguments

...

Additional parameters passed to multiple functions; see details.

data

Name of a package dataset; used by read_sf_pkg() only.

path

A file path.

bbox

A bounding box object; defaults to NULL. If "bbox" is provided, only returns features intersecting the bounding box.

package

Package name; used by read_sf_pkg() only.

filetype

File type supported by sf::read_sf(); Default: 'gpkg'; used by read_sf_pkg() only and required only if the data is in the package cache directory or extdata system files.

dsn

data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder, or contain the name and access credentials of a database); in case of GeoJSON, dsn may be the character string holding the geojson data. It can also be an open database connection.

query

SQL query to select records; see details

name

Name to make file name converted to snake case with janitor::make_clean_names(), e.g. "Residential zoning map" becomes "residential_zoning_map"

wkt_filter

character; WKT representation of a spatial filter (may be used as bounding box, selecting overlapping geometries); see examples

zm_drop

If TRUE, drop Z and/or M dimensions using sf::st_zm

sheet

Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to the first sheet.

coords

Character vector with coordinate values. Coordinates must use the same crs as the from_crs parameter.

from_crs

For df_to_sf, coordinate reference system used by coordinates or well known text in data frame.

geo

If TRUE, use address_to_sf to geocode address column; defaults to FALSE.

address

Address column name, Default: 'address'

url

A url for a spatial data file, tabular data with coordinates, or a ArcGIS FeatureServer or MapServer to access with esri2sf::esri2sf()

show_col_types

If FALSE, do not show the guessed column types. If TRUE always show the column types, even if they are supplied. If NULL (the default) only show the column types if they are not explicitly supplied by the col_types argument.

id

The name of a column in which to store the file path. This is useful when reading multiple input files and there is data in the file paths, such as the data collection date. If NULL (the default) no extra column is created.

filename

File name; if file name is NULL, name and file type are both required.

prefix

File name prefix. "date" adds a date prefix, "time" adds a date/time prefix; defaults to NULL.

method

the geocoding service to be used. API keys are loaded from environmental variables. Run usethis::edit_r_environ() to open your .Renviron file and add an API key as an environmental variable. For example, add the line GEOCODIO_API_KEY="YourAPIKeyHere".

"osm": Nominatim.

"census": US Census. Geographic coverage is limited to the United States. Batch geocoding is supported.

"arcgis": ArcGIS.

"geocodio": Geocodio. Geographic coverage is limited to the United States and Canada. An API key must be stored in the environmental variable "GEOCODIO_API_KEY". Batch geocoding is supported.

"iq": Location IQ. An API key must be stored in the environmental variable "LOCATIONIQ_API_KEY".

"google": Google. An API key must be stored in the environmental variable "GOOGLEGEOCODE_API_KEY".

"opencage": OpenCage. An API key must be stored in the environmental variable "OPENCAGE_KEY".

"mapbox": Mapbox. An API key must be stored in the environmental variable "MAPBOX_API_KEY".

"here": HERE. An API key must be stored in the environmental variable "HERE_API_KEY". Batch geocoding is supported, but must be explicitly called with mode = "batch".

"tomtom": TomTom. An API key must be stored in the environmental variable "TOMTOM_API_KEY". Batch geocoding is supported.

"mapquest": MapQuest. An API key must be stored in the environmental variable "MAPQUEST_API_KEY". Batch geocoding is supported.

"bing": Bing. An API key must be stored in the environmental variable "BINGMAPS_API_KEY". Batch geocoding is supported, but must be explicitly called with mode = "batch".

"geoapify": Geoapify. An API key must be stored in the environmental variable "GEOAPIFY_KEY".

"cascade" use geocode_combine or geo_combine instead. The "cascade" method first uses one geocoding service and then uses a second geocoding service if the first service didn't return results. The services and order is specified by the cascade_order argument. Note that this is not compatible with full_results = TRUE as geocoding services have different columns that they return.

unzip

If TRUE, url must be a zip file that is downloaded, unzipped into a temporary directory (created with tempdir()), and then read to a file using the specified file type.

ss

Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

ask

If TRUE, ask for the name of the Google Sheet to read if ss is not provided to read_sf_gsheet.

Details

Reading data from a url:

read_sf_url supports multiple types of urls:

A MapServer or FeatureServer URL

A URL for a GitHub gist with a single spatial data file (first file used if gist contains multiple)

A URL for a spatial data file or a CSV file

A Google Sheets URL

A public Google Maps URL

Reading data from a package:

read_sf_pkg looks for three types of package data:

Data loaded with the package

External data in the extdata system files folder.

Cached data in the cache directory returned by rappdirs::user_cache_dir

Additional ... parameters:

read_sf_pkg and read_sf_download both pass additional parameters to read_sf_path which supports query, name_col, name, and table. name and name_col are ignored if a query parameter is provided. If table is not provided, a expected layer name is created based on the file path.

read_sf_url pass the where, name_col, and name for any ArcGIS FeatureServer or MapServer url (passed to read_sf_esri) or sheet if the url is for a Google Sheet (passed to googlesheets4::read_sheet), or a query or wkt filter parameter if the url is some other type (passed to sf::read_sf).

read_sf_ext is a flexible function that only has bbox as a named parameter and all other parameters in ... are passed to one of the other overedge read_sf functions.

See also